# gbvmis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a gbvmis api",
        "title": "SGBV Police API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080/api",
    "basePath": "/",
    "paths": {
...
        "/examination": {
            "post": {
                "description": "Creates a new examination entry in the system and returns the created record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinations"
                ],
                "summary": "Create a new examination record",
                "parameters": [
                    {
                        "description": "Examination data to create",
                        "name": "examination",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Examination"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created examination record",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid input",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when creating examination",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/examination/{id}": {
            "get": {
                "description": "Fetches a examination record based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinations"
                ],
                "summary": "Retrieve a single examination record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Examination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Examination retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Examination not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when retrieving examination",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the details of a examination record based on the provided ID and request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinations"
                ],
                "summary": "Update an existing examination record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Examination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Examination data to update",
                        "name": "examination",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateExaminationPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Examination updated successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Invalid input or empty request body",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Examination not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when updating examination",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a examination record based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinations"
                ],
                "summary": "Delete a examination record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Examination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Examination deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Examination not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when deleting examination",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/examinations": {
            "get": {
                "description": "Fetches all examination records with pagination support.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinations"
                ],
                "summary": "Retrieve a paginated list of examinations",
                "responses": {
                    "200": {
                        "description": "Examinations retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve examinations",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/examinations/search": {
            "get": {
                "description": "Retrieves a paginated list of examinations based on search criteria.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Examinations"
                ],
                "summary": "Search for examinations with pagination",
                "responses": {
                    "200": {
                        "description": "Examinations retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve examinations",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.UpdateExaminationPayload": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "integer"
                },
                "consent_given": {
                    "type": "boolean"
                },
                "exam_date": {
                    "type": "string"
                },
                "facility_id": {
                    "type": "integer"
                },
                "findings": {
                    "type": "string"
                },
                "practitioner_id": {
                    "type": "integer"
                },
                "referral": {
                    "description": "Optional referral info",
                    "type": "string"
                },
                "treatment": {
                    "type": "string"
                },
                "victim_id": {
                    "type": "integer"
                }
            }
        },
...
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Case": {
            "type": "object",
            "properties": {
                "case_number": {
                    "type": "string"
                },
                "charges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Charge"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "date_opened": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "officer": {
                    "$ref": "#/definitions/models.PoliceOfficer"
                },
                "officer_id": {
                    "type": "integer"
                },
                "policePost": {
                    "$ref": "#/definitions/models.PolicePost"
                },
                "police_post_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "suspect_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "victims": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Victim"
                    }
                }
            }
        },
        "models.Charge": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "integer"
                },
                "charge_title": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "severity": {
                    "description": "e.g., Felony, Misdemeanor",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Examination": {
            "type": "object",
            "properties": {
                "case": {
                    "$ref": "#/definitions/models.Case"
                },
                "case_id": {
                    "type": "integer"
                },
                "consent_given": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "exam_date": {
                    "type": "string"
                },
                "facility": {
                    "$ref": "#/definitions/models.HealthFacility"
                },
                "facility_id": {
                    "type": "integer"
                },
                "findings": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "practitioner": {
                    "$ref": "#/definitions/models.HealthPractitioner"
                },
                "practitioner_id": {
                    "type": "integer"
                },
                "referral": {
                    "description": "Optional referral info",
                    "type": "string"
                },
                "treatment": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "victim": {
                    "$ref": "#/definitions/models.Victim"
                },
                "victim_id": {
                    "type": "integer"
                }
            }
        },
        "models.HealthFacility": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "practitioners": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.HealthPractitioner"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.HealthPractitioner": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "examinations": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Examination"
                    }
                },
                "facility_id": {
                    "description": "Foreign key to HealthFacility",
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "profession": {
                    "description": "e.g., Doctor, Nurse, Counselor",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.PoliceOfficer": {
            "type": "object",
            "properties": {
                "badge_no": {
                    "type": "string"
                },
                "cases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Case"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "post_id": {
                    "type": "integer"
                },
                "rank": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Role"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.PolicePost": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "officers": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PoliceOfficer"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Victim": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cases": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Case"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "dob": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "nin": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        }
    }
}